refactor: Moved each config provider to own file
file formats are now handled in different source files
